Write model tests
